Refactor theme context: move ThemeProviderContext and related types to a separate file for better organization
